# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SakanMatch API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type SignupRequestRole =
  (typeof SignupRequestRole)[keyof typeof SignupRequestRole];

export const SignupRequestRole = {
  owner: "owner",
  seeker: "seeker",
} as const;

export interface SignupRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name?: string;
  role: SignupRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserResponseRole =
  (typeof UserResponseRole)[keyof typeof UserResponseRole];

export const UserResponseRole = {
  owner: "owner",
  seeker: "seeker",
} as const;

export interface UserResponse {
  id: number;
  email: string;
  name?: string | null;
  role: UserResponseRole;
  isPremium: boolean;
  premiumActivatedAt?: string | null;
  premiumSource?: string | null;
  subscriptionStatus?: string | null;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: UserResponse;
}

export interface CreateListingRequest {
  title: string;
  description?: string;
  price: number;
  city: string;
  images?: string[];
}

export interface ListingResponse {
  id: number;
  title: string;
  description?: string | null;
  price: number;
  city: string;
  images?: string[];
  ownerId: number;
  ownerEmail?: string;
  ownerName?: string | null;
  isFeatured?: boolean;
  viewCount?: number | null;
  contactClickCount?: number | null;
  requestCount?: number | null;
  createdAt?: string;
}

export interface CreateSubscriptionResponse {
  subscriptionID: string;
  approvalUrl?: string | null;
}

export interface ActivateSubscriptionRequest {
  subscriptionID: string;
}

export interface ActivateSubscriptionResponse {
  status: string;
  message?: string | null;
}

export interface PremiumStatusResponse {
  isPremium: boolean;
  premiumActivatedAt?: string | null;
  premiumSource?: string | null;
  subscriptionStatus?: string | null;
  paypalSubscriptionId?: string | null;
  lastPaymentAt?: string | null;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type GetListingsParams = {
  city?: string;
  minPrice?: number;
  maxPrice?: number;
};